+++ model_utils/tests/models.py
@@ -1,4 +1,3 @@
-import django
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
@@ -30,9 +29,8 @@ class InheritanceManagerTestChild1(InheritanceManagerTestParent):
     pass
 
 
-if django.VERSION >= (1, 6, 0):
-    class InheritanceManagerTestGrandChild1(InheritanceManagerTestChild1):
-        text_field = models.TextField()
+class InheritanceManagerTestGrandChild1(InheritanceManagerTestChild1):
+    text_field = models.TextField()
 
 
 class InheritanceManagerTestChild2(InheritanceManagerTestParent):